+++ controllers/admin/role.controller.js
@@ -28,12 +28,18 @@ module.exports.create = async (req, res) => {
 
 // [POST] /admin/roles/create
 module.exports.createRole = async (req, res) => {
-    console.log(req.body);
 
-    const record = new Roles(req.body);
-    await record.save();
+    // console.log(req.body);
+    const permissions= res.locals.role.permissions;
 
-    res.redirect(`${systemConfig.prefixAdmin}/roles`);
+    if(permissions.includes("roles_create")){
+        const record = new Roles(req.body);
+        await record.save();
+
+        res.redirect(`${systemConfig.prefixAdmin}/roles`);
+    }else{
+        
+    }
 };
 
 // [GET] /admin/roles/edit/:id
@@ -62,20 +68,25 @@ module.exports.edit = async (req, res) => {
 // [PATCH] /admin/roles/edit/:id
 module.exports.editRole = async (req, res) => {
 
-    try{
-        const id = req.params.id;
+    const permissions= res.locals.role.permissions;
 
-        await Roles.updateOne({_id:id}, req.body);
-
-        req.flash("success", "Cập nhập nhóm quyền thành công!");
-
-        res.redirect("back");
-    }catch{
-        req.flash("error", "Cập nhập nhóm quyền thất bại!");
-
-        res.redirect("back");
-    }
+    if(permissions.includes("roles_edit")){
+        try{
+            const id = req.params.id;
     
+            await Roles.updateOne({_id:id}, req.body);
+    
+            req.flash("success", "Cập nhập nhóm quyền thành công!");
+    
+            res.redirect("back");
+        }catch{
+            req.flash("error", "Cập nhập nhóm quyền thất bại!");
+    
+            res.redirect("back");
+        }
+    }else{
+        
+    }
 };
 
 // [GET] /admin/roles/permissions
@@ -95,44 +106,54 @@ module.exports.permissions = async (req, res) => {
 // [PATCH] /admin/roles/permission
 module.exports.permissionsPatch = async (req, res) => {
     // console.log(req.body.permissions);
+
+    const permissions= res.locals.role.permissions;
+
+    if(permissions.includes("roles_permissions")){
+        try{
+            const permissions = JSON.parse(req.body.permissions);
+            for(const item of permissions){
+                await Roles.updateOne({_id:item.id}, {permissions: item.permissions});
+            }
+            // console.log(permissions);
     
-    try{
-        const permissions = JSON.parse(req.body.permissions);
-        for(const item of permissions){
-            await Roles.updateOne({_id:item.id}, {permissions: item.permissions});
+            req.flash("success", "Cập nhập thành công");
+    
+            res.redirect("back");
+        }catch{
+            req.flash("error", "Cập nhập thất bại");
+    
+            res.redirect("back");
         }
-        // console.log(permissions);
-
-        req.flash("success", "Cập nhập thành công");
-
-        res.redirect("back");
-    }catch{
-        req.flash("error", "Cập nhập thất bại");
-
-        res.redirect("back");
+    }else{
+        
     }
-    
 };
 
 // [DELETE] /admin/roles/delete/:id
 module.exports.deleteItem = async (req, res) => {
-    console.log(req.params.id);
+    // console.log(req.params.id);
 
-    try{
-        await Roles.updateOne({_id: req.params.id}, {
-            deleted: true,
-            deletedBy: {
-                account_id: res.locals.user.id,
-                deletedAt: new Date()
-            }
-        });
-        req.flash("success", "Xoá sản phẩm thành công!");
-        res.redirect("back");
-    }catch{
-        req.flash("error", "Xoá sản phẩm thất bại!");
-        res.redirect("back");
+    const permissions= res.locals.role.permissions;
+
+    if(permissions.includes("roles_delete")){
+        try{
+            await Roles.updateOne({_id: req.params.id}, {
+                deleted: true,
+                deletedBy: {
+                    account_id: res.locals.user.id,
+                    deletedAt: new Date()
+                }
+            });
+            req.flash("success", "Xoá sản phẩm thành công!");
+            res.redirect("back");
+        }catch{
+            req.flash("error", "Xoá sản phẩm thất bại!");
+            res.redirect("back");
+        }
+    }else{
+        
     }
-
 };
 
 // [GET] /admin/roles/bin
@@ -153,35 +174,42 @@ module.exports.bin = async (req, res) => {
 // [PATCH] /admin/roles/bin/restore/:id
 module.exports.restoreItem= async (req, res) => {
     // console.log(req.params.id);
+    const permissions= res.locals.role.permissions;
 
-    try{
-        await Roles.updateOne(
-            {_id: req.params.id},
-            {deleted: false}
-        );
-    
-        req.flash("success", "Khôi phục nhóm quyền thành công!");
-        res.redirect("back");
-    }catch{
-        req.flash("error", "Khôi phục nhóm quyền thất bại!");
-        res.redirect("back");
+    if(permissions.includes("roles_restore")){
+        try{
+            await Roles.updateOne(
+                {_id: req.params.id},
+                {deleted: false}
+            );
+        
+            req.flash("success", "Khôi phục nhóm quyền thành công!");
+            res.redirect("back");
+        }catch{
+            req.flash("error", "Khôi phục nhóm quyền thất bại!");
+            res.redirect("back");
+        }
+    }else{
+        
     }
-
-    
 };
 
 // [DELETE] /admin/roles/bin/delete/:id
 module.exports.deleteItemBin= async (req, res) => {
     // console.log(req.params.id);
+    const permissions= res.locals.role.permissions;
 
-    try{
-        await Roles.deleteOne({_id: req.params.id});
-
-        req.flash("success", "Xoá hẳn nhóm quyền thành công!");
-        res.redirect("back");
-    }catch{
-        req.flash("error", "Xoá hẳn nhóm quyền thất bại!");
-        res.redirect("back");
-    }
+    if(permissions.includes("roles-bin_delete")){
+        try{
+            await Roles.deleteOne({_id: req.params.id});
     
-}
+            req.flash("success", "Xoá hẳn nhóm quyền thành công!");
+            res.redirect("back");
+        }catch{
+            req.flash("error", "Xoá hẳn nhóm quyền thất bại!");
+            res.redirect("back");
+        }
+    }else{
+        
+    }
+};
